Updated ArrayUtils.java
Added a couple more methods and commented (most of) the methods above the 3 lines of "////////////" (line 405). If necessary, methods to fit the list structures from the project will be implemented. The goal of ArrayUtils.java is not to have it tailored for the project (e.g. putting a "printMarchMadnessResults(...)" method in this file), but instead, have the functions usable for the necessary portions of the project.

The code compiles, and most functions were tested briefly, but not extensively.