Started KNN "Pseudocode"/Wrote ideas in comments
I tried to simplify some of the pseudocode I found on other sites. I'm
not sure if the argmin function will actually work or not based on the
pseudocode that's there. I may have to write out some test data and run
it through the algorithm to make sure it works. With the ideas I have in
place, there shouldn't be much trouble returning the tuples stating
which team wins each game. Clustering may require limits on cluster
sizes (We don't want a single team having their own cluster. It won't
say anything useful), and this will have to be implemented carefully.
Infinite loops could occur if we move teams between clusters if it's
implemented incorrectly (i.e. a tiebreaker could possibly cause a team
to move from cluster 1 to 2, to 1, to 2, to 1 etc...).

Finding out where the clusters are stored may be a design choice (stored
in Prototypes themselves, or stored in separate class etc.). I may do
something similar to the Agglomeration class, and I may move the
internal Prototype class outside of the Agglomeration class.