Ignore delclasses.bat
Sometimes, recompiling wasn't updating a few of the class files for me
(there are reasons why javac doesn't do this properly sometimes). I'd
run this batch script to delete all .class files and then perform a
fresh recompilation to guarantee a clean and error-free compile that was
free of "NoClassDefFoundError" runtime errors.